+++ Assets/Scripts/Currency/CurrencyDrop.cs
@@ -17,7 +17,7 @@ public class CurrencyDrop : RecycleObject
     /// <summary>
     /// 드랍되는 재화 양
     /// </summary>
-    private uint amount = 1;
+    private ulong amount = 1;
 
     /// <summary>
     /// 오브젝트가 사라지기까지 시간
@@ -71,7 +71,7 @@ public class CurrencyDrop : RecycleObject
     /// <summary>
     /// 외부(몬스터)에서 드랍 타입과 양을 세팅해 주는 메서드
     /// </summary>
-    public void SetDropInfo(uint dropAmount)
+    public void SetDropInfo(ulong dropAmount)
     {
         amount = dropAmount;
     }
@@ -118,7 +118,7 @@ public class CurrencyDrop : RecycleObject
                 UserDataManager.Instance.AddCurrency_Gem(amount);
                 break;
             case CurrencyType.Diamond:
-                UserDataManager.Instance.AddCurrency_Diamond(amount);
+                UserDataManager.Instance.AddCurrency_Diamond((uint)amount);
                 break;
         }
     }
