+++ Assets/Scripts/Manager/UserDataManager.cs
@@ -120,7 +120,7 @@ public class UserDataManager : Singleton<UserDataManager>
     /// <summary>
     /// Gold 추가 함수
     /// </summary>
-    public void AddCurrency_Gold(uint amount)
+    public void AddCurrency_Gold(ulong amount)
     {
         if (amount <= 0) return;
         Currency_Gold += amount;
@@ -131,7 +131,7 @@ public class UserDataManager : Singleton<UserDataManager>
     /// <summary>
     /// Gold 사용(차감) 함수
     /// </summary>
-    public bool ConsumeCurrency_Gold(uint amount)
+    public bool ConsumeCurrency_Gold(ulong amount)
     {
         if (amount <= 0 || amount > Currency_Gold)
             return false;
@@ -146,7 +146,7 @@ public class UserDataManager : Singleton<UserDataManager>
     /// <summary>
     /// Gold 추가 함수
     /// </summary>
-    public void AddCurrency_Gem(uint amount)
+    public void AddCurrency_Gem(ulong amount)
     {
         if (amount <= 0) return;
         Currency_Gem += amount;
@@ -157,7 +157,7 @@ public class UserDataManager : Singleton<UserDataManager>
     /// <summary>
     /// Gold 사용(차감) 함수
     /// </summary>
-    public bool ConsumeCurrency_Gem(uint amount)
+    public bool ConsumeCurrency_Gem(ulong amount)
     {
         if (amount <= 0 || amount > Currency_Gem)
             return false;
